Closes #17, all model classes in `architectures.py`, deleted `networks` directory
